Improve security around the CSR/CSR Attributes files and add trusted facts.
Remove the /etc/puppet/csr_attributes.yaml file after its been consumed.
Strictly speaking, the CSR file itself still contains these bits. In the
future we may purge this file as well.

Additionally, add support for 'trusted facts' to be embedded into the
csr_attributes file. While in here, I re-worked the code to simply dump
a hash in YAML form into the file. This makes it simpler in the future
to add more keys or change the keys around.